서버에서 send 오류나던부분 수정
- 서버에서 처리한후 클라로 보내는과정에서 전송 실패되는 부분 수정 완료

+++ Graduation_Server/Graduation_Server/User_Client.h
@@ -8,8 +8,8 @@ private:
 	char		_name[MAX_NAME_SIZE];
 	STATE		_state;
 
-	int			_prev_size;
-	int			_id;
+	unsigned short	_prev_size;
+	int				_id;
 
 public:
 	unordered_set <int> room_list;
